cld: change descriptor pool size variable to u_int16_t
Currently the txrx module's tx descriptor pool size
is stored as an int. This changes it to u_int16_t, to make it
simpler for static analysis tools to confirm that a computed
tx descriptor ID fits into a u_int16_t variable.

CL Id: 2037642

+++ cld/CORE/COMMON/ol_cfg.h
@@ -231,7 +231,7 @@ int ol_cfg_tx_free_at_download(ol_pdev_handle pdev);
  * @param pdev - handle to the physical device
  * @return the number of tx buffers available in a HL target
  */
-int ol_cfg_target_tx_credit(ol_pdev_handle pdev);
+u_int16_t ol_cfg_target_tx_credit(ol_pdev_handle pdev);
 
 
 /**
